requirements.txt
loguru~=0.5.3
bs4~=0.0.1
pyyaml~=5.4.1

requests==2.25.1
beautifulsoup4~=4.9.3
twilio~=6.54.0

lxml
geopy~=2.1.0
cachetools~=4.2.1
pytest~=6.2.3